feature:deleted comment done

+++ src/components/Blog/BlogDetail/CommentList/CommentItem/index.js
@@ -4,6 +4,7 @@ import CreateReply from '../CreateReply'
 import { Link } from 'react-router-dom'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { useDeleteCommentMutation } from '../../../../../services/commentApi'
+import Swal from 'sweetalert2'
 
 const CommentItem = ({
   commentId,
@@ -13,10 +14,46 @@ const CommentItem = ({
   filterReply,
 }) => {
   const [deleteComment] = useDeleteCommentMutation()
+
+  const swalWithBootstrapButtons = Swal.mixin({
+    customClass: {
+      confirmButton: 'btn btn-success',
+      cancelButton: 'btn btn-danger',
+    },
+    buttonsStyling: false,
+  })
+
   const handleDelete = async () => {
     try {
-      console.log('12314')
-      deleteComment({ commentId })
+      await swalWithBootstrapButtons
+        .fire({
+          title: 'Are you sure?',
+          text: "You won't be able to revert this!",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonText: 'Yes, delete it!',
+          cancelButtonText: 'No, cancel!',
+          reverseButtons: true,
+        })
+        .then((result) => {
+          if (result.isConfirmed) {
+            swalWithBootstrapButtons.fire(
+              'Deleted!',
+              'Your file has been deleted.',
+              'success'
+            )
+            deleteComment({ commentId })
+          } else if (
+            /* Read more about handling dismissals below */
+            result.dismiss === Swal.DismissReason.cancel
+          ) {
+            swalWithBootstrapButtons.fire(
+              'Cancelled',
+              'Your imaginary file is safe :)',
+              'error'
+            )
+          }
+        })
     } catch (e) {
       console.log(e)
     }
